Time: 8 ms (71.06%), Space: 29.1 MB (29.88%) - LeetHub

+++ 3296-minimum-number-of-seconds-to-make-mountain-height-zero/3296-minimum-number-of-seconds-to-make-mountain-height-zero.cpp
@@ -22,7 +22,7 @@ public:
         int maxTime = *max_element(begin(workerTimes), end(workerTimes));
         ll l = 1;
         ll r = (ll)maxTime * mountainHeight * (mountainHeight+1)/2;
-        int result  = 0;
+        ll result  = 0;
 
         // khandani template of binary search
         while(l <= r){
